WIP: Add a custom annotation processor to detect whether Eclipse built things
WIP because we still need to run mvn license:update-file-header

We know that Eclipse does a bad job at conforming to the standard required
by Java: it does not execute the annotation processors present on the
class path by default.

The sole purpose of this component is to detect whether it was Eclipse
that built stuff in a certain class path component so we can try to figure
out whether sezpoz has had a chance or not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>