No longer tracking OptionsService
There seemed to be a bug that resulted in occasional NPEs when
double clicking the status bar. Apparently there was a race
condition initializing the OptionsService. Sometimes it would
get assigned null.

Now we are always getting the current OptionsService when mouse
is clicked. This deferrment avoids the NPE.

Going forward it might make sense to fold OptionsService into
the constructor for UIService. If it's handled this way it
would never be null.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5170 be23795f-a60e-40d4-ba19-055798872e57