Initial commit of the ImageJ launcher
Using the NAR plugin, we compile the ImageJ launcher now. Note that it
cannot run in-place because it is neither finding ij-launcher.jar nor the
ImageJ root directory.

Based on the ImageJ.c in fiji.git's 'master' branch as of Mon Mar 26
23:26:45 2012 -0500.

The launcher module requires a C compiler (gcc, to be precise), so it is
not included by default; Activate the 'launcher' profile to compile it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5195 be23795f-a60e-40d4-ba19-055798872e57