Updater UpToDateCheck: add a mode for running from within Eclipse
Since Eclipse cannot build .jar files automatically, it does not
make sense to run the Updater upon startup when ImageJ was launched
from within Eclipse. Infer this situation from looking whether the
UpToDate class lives inside a .jar file or not, and just ignore the
up-to-date check altoghether if it does not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5234 be23795f-a60e-40d4-ba19-055798872e57