Fix profile issue with Swing UI
The problem is that the new "not-eclipse" workaround profile causes the
swing profile's "activeByDefault" to no longer be activated.

We fix the issue by activating all UI profiles based on an "ij.ui"
property. If it is not set, Swing is used; otherwise, the value
indicates the desired UI.

Running "mvn -P all" will still build everything. :-)

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5244 be23795f-a60e-40d4-ba19-055798872e57