Updater: Prepare the dependency checking for version numbers in .jar names
We decided that it would be better for ImageJ2 to keep the version numbers
in the file names.

As a consequence, the Updater now has to learn to deal with the situation.

As a first step, the dependency analyzer no longer bothers with exact file
names but only with the initial part of the file name.

The next steps are more crucial: FileObject will need to learn to look
harder for dependencies (and to be gentle when the file name does not
match exactly, but allow for version differences; I am thinking about a
regex like this: ^(.*)-(\d+\.\d(\.\d)-.*)\.jar$).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5266 be23795f-a60e-40d4-ba19-055798872e57