Re-use the ImageJ launcher for Fiji
For the short term, the ImageJ launcher needs to call fiji.Main (which
calls legacy ImageJ1) for the moment. Detect that we're running inside
Fiji by looking for jars/fiji-compat.jar and do not call imagej.Main
in that case.

In the medium term, we will switch the default to imagej.Main.

The newly introduced ImageJ.cfg setting legacy.mode was introduced for
a future Edit>Option>Legacy Mode or even Help>Launch Legacy ImageJ
(although the latter might be done by enhancing ClassLoaderPlus
slightly).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5314 be23795f-a60e-40d4-ba19-055798872e57