Improve display activation model
Changed activation model so that you ask the display service to activate
and it manages the activation. Because of this,
DisplayService.getActiveDisplay is definitive - the value returned is
the target active display, even if several display activation events are
in the queue. Also changed the Swing user interface service to respond
to an activation only after any current events have been processed and
to do so on the currently active display. The UI synchs to the
activation model after all user interactions have been digested.

Authored-by: Lee Kamentsky

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5324 be23795f-a60e-40d4-ba19-055798872e57