Propagate DisplayUpdatedEvents in a better way
DisplayUpdatedEvent was queued before DisplayCreatedEvent was processed
by UI. JHotDrawImageCanvas was registering for DisplayUpdatedEvent after
it was published, so it never got the first one. I've changed it so that
the Swing user interface's DisplayUpdatedEvent handler propagates the
event downward through the displayViewer, pane and canvas and that's
registered at startup.

Authored-by: Lee Kamentsky

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5328 be23795f-a60e-40d4-ba19-055798872e57