Math plugins now respect region shape
The math plugins have been updated to respect the shape of the currently
selected region. If no selection exists then they are bounded by the current
plane. The operations can be extended to all planes in the image.

To implement this the AbstractAssignPlugin no longer relies on the
ImageAssignment class. Instead it relies on a PointSet approach. The
resulting code is now simpler but single threaded.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@5388 be23795f-a60e-40d4-ba19-055798872e57